pyproject.toml
[project]
name = "personas"
version = "0.1.0"
description = "Persona generation pipeline with behavioral & linguistic profile extraction and simulated patient agents."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Personas Project" }
]
dependencies = [
  "litellm>=1.51.0",
  "pydantic>=2.9.0",
  "typer>=0.12.0",
  "rich>=13.9.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "dspy-ai>=2.4.0",
]




